% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_parts}
\alias{collect_parts}
\title{Collect and combine intronic/exonic parts by gene}
\usage{
collect_parts(annotation, ...)
}
\arguments{
\item{annotation}{an object representing an annotation such as an
`GenomicFeatures::TxDb`, a `GenomicRanges::GRanges` object obtained 
from `rtracklayer::import()` or `plyranges::read_gff()` or the name of
GFF/GTF file.}

\item{...}{optional arguments to be passed to downstream methods}
}
\value{
a GRanges object
}
\description{
Collect and combine intronic/exonic parts by gene
}
\details{
This function takes an annotation object and returns
a GRanges object with number of rows equal to the number of gene 
features contained. The resulting GRanges object has the following 
additional metadata columns:

* `exonic_parts`: a GRangesList column containing the exonic regions 
for each gene, obtained from `type == 'exon'` in GFF/GTF file. In the 
case of a TxDb object, this is obtained by `GenomicFeatures::exonBy`.
* `intronic_parts`: a GRangesList column containing the intronic regions
for each gene. This is simply defined as the parallel set difference betweeen
the gene ranges and the exonic ranges.
* `n_olaps`: the integer count of the number times a gene overlaps
any other gene.
}
\examples{
gtf <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf", package = "airway")
collect_parts(gtf)

}
