% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{flatten_merge}
\alias{flatten_merge}
\title{Flatten exonic/intronic parts and intersect with ranges}
\usage{
flatten_merge(x, parts)
}
\arguments{
\item{x}{a GRanges object}

\item{parts}{an annotation GRanges from \code{collect_parts()}}
}
\description{
Flatten exonic/intronic parts and intersect with ranges
}
