% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather_coverage}
\alias{gather_coverage}
\title{Generate long-form coverage as coverage}
\usage{
gather_coverage(bams, genome_info, BPARAM = BiocParallel::bpparam())
}
\arguments{
\item{bams}{a character vector containing a list of paths to indexed BAM files}

\item{genome_info}{a GRanges object containing reference annotation}

\item{BPARAM}{a BiocParallel object (default = `bpparam()`)}
}
\description{
Generate long-form coverage as coverage
}
