% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-long.R
\name{join_design}
\alias{join_design}
\title{Merge a design matrix onto a GRanges object}
\usage{
join_design(ranges, design, by = NULL)
}
\arguments{
\item{ranges}{a GRanges object}

\item{design}{a DataFrame object with}

\item{by}{the key column either a character vector of the common columns
between `ranges` and `design` or a named character vector (default = NULL).}
}
\description{
Merge a design matrix onto a GRanges object
}
\details{
To summarise ranges features over variables in the experimental `design`,
we can join the `design`` matrix to a `ranges` object. 
This function computes an inner join on the common key between the two 
objects and sorts by the key.
}
\seealso{
`dplyr::inner_join()`
}
