% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{merge_coverage}
\alias{merge_coverage}
\title{Coverage over features}
\usage{
merge_coverage(cvg, features)
}
\arguments{
\item{cvg}{a GRanges object from `gather_coverage()`}

\item{features}{a GRanges object from `prepare_annotation()`}
}
\description{
Coverage over features
}
