% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{merge_design}
\alias{merge_design}
\title{Merge experimental design onto coverage features}
\usage{
merge_design(cvg, design, on = NULL)
}
\arguments{
\item{cvg}{a GRanges object from `gather_coverage()` or `merge_coverage()`}

\item{design}{a DataFrame object with}

\item{on}{a column name in `design` that links to the `source` column in `cvg`.}
}
\description{
Merge experimental design onto coverage features
}
\details{
To summarise coverage features over variables in the experimental `design`,
we require those features to be joins to the coverage GRanges. This
requires the design table to have a column that matches the file names
generated by `gather_coverage()`.
}
