% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{nest_by_overlaps}
\alias{nest_by_overlaps}
\title{Nest by Overlapping Ranges}
\usage{
nest_by_overlaps(ranges, features, .key, .index)
}
\arguments{
\item{ranges}{a GRanges object}

\item{features}{a GRanges object from `collect_parts()`}

\item{key}{one or more columns to group ranges}

\item{index}{one or more columns to group overlaps by}
}
\value{
a GRanges object
}
\description{
Nest by Overlapping Ranges
}
\details{
This function restricts the coverage GRanges, `cvg` to the ranges
that intersect the `exonic_parts` and `intronic_parts` ranges, the result
is a GRanges object with additional columns: 

* `gene_id`, the gene_id from `features` corresponding to an exon/intron.
* `feature_type`, whether the range corresponds to an exon or intro feature.
* `feature_rank`, the rank of the exon/intron feature within a gene.
* `feature_length`, the width of the exon/intron
* `feature_strand`, the strand of the exon/intron
}
