% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{plot_cvg_over_gene}
\alias{plot_cvg_over_gene}
\title{Visualisation Helpers}
\usage{
plot_cvg_over_gene(x, gene)
}
\arguments{
\item{x}{a GRanges object obtained from `merge_coverage()`}

\item{gene}{a GRanges object from `prepare_annotation()` containing a gene_id column}
}
\description{
Visualisation Helpers
}
