% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_annotation}
\alias{prepare_annotation}
\title{GTF processing:}
\usage{
prepare_annotation(.data)
}
\arguments{
\item{.data}{a GRanges object obtained from `rtracklayer::import()` or
`plyranges::read_gff()`.}
}
\value{
a GRanges object
}
\description{
1. Import the GTF file and extract genes and exon features. Retain any relevant annotation
information.
2. Mark genes if they self-overlap.
3. Split each gene into exon/intron coordinates. An intron is any region within
a gene not annotated as an exon.
4. Report total exon and intron lengths in each gene, where exons are the
union set of all transcripts from the same gene.
Prepare GTF/GFF file for coverage analysis
}
\details{
This function takes a GFF/GTF file as GRanges object and returns
a GRanges object with number of rows equal to the number of gene 
features contained. The result has  with the following metadata columns:

* `gene_id, gene_type, gene_name`: these are kept from the input `.data`
* `simple_exonic`: a GRangesList column containing the exonic regions 
for each gene, obtained from `type == 'exon'` in the GFF/GTF GRanges object.
* `simple_intronic`: a GRangesList column containing the intronic regions
for each gene. This is simply defined as the set difference betweeen
the gene ranges and the exonic ranges.
* `n_olaps`: the integer count of the number times a gene overlaps
any other gene.
}
