% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processAnnoDb}
\alias{processAnnoDb}
\title{Internal generic to process a GFF/GTF file}
\usage{
processAnnoDb(db)
}
\description{
GTF processing:
}
\details{
1. Import the GTF file and extract genes and exon features. Retain any relevant annotation
information.
2. Mark genes if they self-overlap.
3. Split each gene into exon/intron coordinates. An intron is any region within
a gene not annotated as an exon.
4. Report total exon and intron lengths in each gene, where exons are the
union set of all transcripts from the same gene.
}
