% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangenostics.R
\docType{methods}
\name{rangle}
\alias{rangle}
\alias{rangle,GenomicRanges-method}
\alias{rangle,GroupedGenomicRanges-method}
\title{Compute diagnostics over Ranges}
\usage{
rangle(x, .var, .funs, ...)

\S4method{rangle}{GenomicRanges}(x, .var, .funs, ...)

\S4method{rangle}{GroupedGenomicRanges}(x, .var, .funs, ...)
}
\arguments{
\item{x}{a Ranges object}

\item{.var}{A meta data column in \code{x} to summarise over either a bare variable name
or a character vector of length 1.}

\item{.funs}{A list of functions to summarise}

\item{...}{additional parameters to pass to \code{.funs}}
}
\description{
Compute diagnostics over Ranges
}
