% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangenostics.R
\name{rangenostics}
\alias{rangenostics}
\title{Compute diagnoistics over Ranges (rangenostics)}
\usage{
rangenostics(ranges, .var, .funs, ...)
}
\arguments{
\item{ranges}{a GRanges object}

\item{.var}{A column in \code{ranges} defining an index}

\item{.funs}{A list of functions to summarise}

\item{...}{additional parameters to pass to ...}
}
\description{
Compute diagnoistics over Ranges (rangenostics)
}
