% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangnostics.R
\name{rangnostics}
\alias{rangnostics}
\title{Compute diagnoistics over Ranges (rangnostics)}
\usage{
rangnostics(ranges, .var, .funs, ...)
}
\arguments{
\item{ranges}{a GRanges object}

\item{.var}{A column in \code{ranges} defining an index}

\item{.funs}{A list of functions to summarise}

\item{...}{additional parameters to pass to ...}
}
\description{
Compute diagnoistics over Ranges (rangnostics)
}
