% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangenostics.R
\name{rango}
\alias{rango}
\title{Compute diagnoistics over Ranges (rango)}
\usage{
rango(x, .var, .funs, ...)
}
\arguments{
\item{x}{a Ranges object}

\item{.var}{A meta data column in \code{x} to summarise over}

\item{.funs}{A list of functions to summarise}

\item{...}{additional parameters to pass to \code{.funs}}
}
\description{
Compute diagnoistics over Ranges (rango)
}
