% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangle.R
\name{rng_vars}
\alias{rng_vars}
\title{Select range variables}
\usage{
rng_vars(...)
}
\value{
A list of quosures
}
\description{
This helper function is used to provide semantics
for selecting variables. It is useful for specifying the index
in \code{rangle()} or passing to scoped variants in \code{dplyr} like
\code{dplyr::mutate_at()}.
}
