% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-split.R
\docType{methods}
\name{split_ranges}
\alias{split_ranges}
\alias{split_ranges,Ranges-method}
\alias{split_ranges,GroupedGenomicRanges-method}
\alias{split_ranges,GroupedIntegerRanges-method}
\title{Cast a Ranges to a RangesList}
\usage{
split_ranges(x, ...)

\S4method{split_ranges}{Ranges}(x, ...)

\S4method{split_ranges}{GroupedGenomicRanges}(x, ...)

\S4method{split_ranges}{GroupedIntegerRanges}(x, ...)
}
\arguments{
\item{x}{a Ranges object}

\item{...}{grouping variables to split by (passed to \code{plyranges::group_by()})}
}
\value{
a RangesList object
}
\description{
Cast a Ranges to a RangesList
}
