% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{squish_coverage}
\alias{squish_coverage}
\title{Squish coverage over intronic/exonic parts}
\usage{
squish_coverage(x, ..., threshold = 0L)
}
\arguments{
\item{x}{a GRanges object from `merge_coverage()\merge_design()`}

\item{...}{columns to partition x}

\item{threshold}{Filter bases covered to above some integer threshold (default = 0L).}
}
\value{
a GRanges object, where each gene is split into exonic
and intronic parts, and the coverage is summarised as follows over each
exon / intron:

* `gene_id` the gene_id as from `prepare_annotation()`
* `feature_type` whether a feature is an intron or exon
* `feature_length` the total length of the intronic/exonic part
* `total_coverage`
* `average_coverage`
* `n_bases_above_threshold`
* `prop_bases_covered_at_threshold`
}
\description{
Squish coverage over intronic/exonic parts
}
\details{
Here we use `plyranges::reduce_ranges()` to squish
the coverage GRanges to have coverage over each exon/intron within
a gene. If `...` are supplied additional grouping is added by metadata
columns in x, using `plyranges::group_by()`.
}
