% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{standardise_coverage}
\alias{standardise_coverage}
\title{A wrapper to compute coverage but only on 'standard' chromosomes}
\usage{
standardise_coverage(x, genome_info)
}
\description{
A wrapper to compute coverage but only on 'standard' chromosomes
}
