% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{summarise_intronic}
\alias{summarise_intronic}
\title{Intron Signal Aggregator}
\usage{
summarise_intronic(x, ...)
}
\arguments{
\item{x}{a GRanges object from `merge_coverage()\merge_design()`}

\item{...}{columns to partition x}
}
\value{
a GRanges object with the following metadata columns
* GeneId
* IsExpressed: a flag for whether gene is expressed or not
* ExonLength
* IntronLength
* MeanExonCoverage
* MeanIntronCoverage
}
\description{
Intron Signal Aggregator
}
\details{
BAM processing
1. Over each gene compute the log2 trimmed mean for exon and intron counts
and flag whether the gene is expressed (>= 3 reads in both exon and intron)
2. Compute the run length of intron ranges exceeding a nominal threshold
}
