% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superintronic-package.R
\docType{package}
\name{superintronic-package}
\alias{superintronic}
\alias{superintronic-package}
\title{superintronic: Tools for Working with Intron Signal}
\description{
superintronic stimates coverage over intronic/exonic regions within
    genes given an annotation and an experimental design. It allows the 
    user to quickly summarise any intron signal, and visualise the
    coverage regions with their gene context.
}
\author{
\strong{Maintainer}: Stuart Lee \email{lee.s@wehi.edu.au}

Authors:
\itemize{
  \item Charity Law \email{law@wehi.edu.au}
}

}
\keyword{internal}
