% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{unnest_parts}
\alias{unnest_parts}
\title{Expand the exonic/intronic parts of a gene into a long GRanges object}
\usage{
unnest_parts(x)
}
\arguments{
\item{x}{a GRanges object from \code{collect_parts()}}
}
\value{
a GRanges object
}
\description{
Expand the exonic/intronic parts of a gene into a long GRanges object
}
\details{
This function restricts the expands/unnests, the ranges from
\code{collect_parts} the result is a GRanges object with the following columns:
\itemize{
\item \code{gene_id}, the gene_id from \code{features} corresponding to an exon/intron.
\item \code{feature_type}, whether the range corresponds to an exon or intro feature.
\item \code{feature_rank}, the rank of the exon/intron feature within a gene.
\item \code{feature_length}, the width of the exon/intron
\item \code{feature_strand}, the strand of the exon/intron
}
}
