% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_coverage}
\alias{view_coverage}
\title{View coverage as a stacked area chart}
\usage{
view_coverage(cvg, target, hline = NULL, facets = ggplot2::vars())
}
\arguments{
\item{cvg}{a GRanges object obtained from `merge_coverage()`}

\item{target}{a GRanges object from `prepare_annotation()` containing a gene_id column}

\item{facets}{a set of variables quoted by `ggplot2::vars()` that is passed 
down to `facet_wrap()`}
}
\description{
View coverage as a stacked area chart
}
