% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_coverage_within_gene}
\alias{view_coverage_within_gene}
\title{Visualisation Helpers}
\usage{
view_coverage_within_gene(cvg, target, facets = ggplot2::vars())
}
\arguments{
\item{cvg}{a GRanges object obtained from `merge_coverage()`}

\item{target}{a GRanges object from `prepare_annotation()` containing a gene_id column}
}
\description{
Visualisation Helpers
}
