% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_exin}
\alias{view_exin}
\title{Scatter exon-intron signal}
\usage{
view_exin(data, x, y, facets = ggplot2::vars(), geom = "point")
}
\arguments{
\item{data}{a DataFrame object obtained from `spread_coverage()`}

\item{x, y}{bare variable names for the x,y plot aesthetics}

\item{facets}{a set of variables quoted by `ggplot2::vars()` that is passed 
down to `facet_wrap()`}

\item{geom}{produce a 'point' scatter or a 'hex' scatter or c('hex', 'point')
for both}
}
\description{
Scatter exon-intron signal
}
