% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-annotation.R
\name{view_segments}
\alias{view_segments}
\title{An opinionated way of plotting intron/exon features}
\usage{
view_segments(data, colour)
}
\arguments{
\item{data}{a GRanges object from \code{collect_parts()}}

\item{colour}{an optional expression or bare variable in \code{data} that represents a feature}
}
\value{
a ggplot object
}
\description{
An opinionated way of plotting intron/exon features
}
