% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-xy.R
\name{view_xy}
\alias{view_xy}
\title{Generate a scatter (xy) plot over a Ranges or DataFrame}
\usage{
view_xy(data, x, y, facets = ggplot2::vars(), geom = "point")
}
\arguments{
\item{data}{a Ranges or DataFrame object obtained}

\item{x, y}{bare variable names for the x,y plot aesthetics}

\item{facets}{a set of variables quoted by \code{ggplot2::vars()} that is passed
down to \code{facet_wrap()}}

\item{geom}{produce a 'point' scatter or a 'hex' scatter or c('hex', 'point')
for both}
}
\value{
a ggplot object
}
\description{
Generate a scatter (xy) plot over a Ranges or DataFrame
}
