% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{tile_map}
\alias{tile_map}
\alias{roll_map}
\alias{stretch_map}
\title{Map a function over windows}
\usage{
tile_map(.x, .size, .fun, ...)

roll_map(x, .size, .step, .fun, ...)

stretch_map(.x, .start, .step, .fun, ...)
}
\arguments{
\item{.x}{An atomic vector or Rle object.}

\item{.size}{The (integer) size of the window.}

\item{.fun}{A function}

\item{...}{Additional arguments passed on to the mapped function}

\item{.step}{The (integer) amount to shift the start of the window by.}

\item{.start}{The (integer) starting position of the window}
}
\description{
Map a function over windows
}
\details{
The map functions apply a function over non-overlapping windows \code{[tile_map()]},
overlapping windows \code{[roll_map()]}, and windows with a fixed start but
increasing width \code{\link[=stretch_map]{stretch_map()}}
}
\examples{
tile_map(1:5, .size = 2, mean)
roll_map(1:5, .size = 2, .step = 1, mean)
stretch_map(1:5, .size = 1, .size = 2, mean)

}
