% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_parts}
\alias{collect_parts}
\alias{collect_parts,GRanges-method}
\alias{collect_parts,character-method}
\alias{collect_parts,GFFFile-method}
\alias{collect_parts,TxDb-method}
\title{Collect and combine intronic/exonic parts by gene}
\usage{
collect_parts(annotation, ...)

\S4method{collect_parts}{GRanges}(annotation, ...)

\S4method{collect_parts}{character}(annotation, ...)

\S4method{collect_parts}{GFFFile}(annotation, ...)

\S4method{collect_parts}{TxDb}(annotation)
}
\arguments{
\item{annotation}{an object representing an annotation such as an
\code{GenomicFeatures::TxDb}, a \code{GenomicRanges::GRanges} object obtained
from \code{rtracklayer::import()} or \code{plyranges::read_gff()} or the name of
GFF/GTF file.}

\item{...}{optional arguments to be passed to downstream methods}
}
\value{
a GRanges object
}
\description{
Collect and combine intronic/exonic parts by gene
}
\details{
This function takes an annotation object and returns
a GRanges object with number of rows equal to the number of gene
features contained. The resulting GRanges object has the following
additional metadata columns:
\itemize{
\item \code{exonic_parts}: a GRangesList column containing the exonic regions
for each gene, obtained from \code{type == 'exon'} in GFF/GTF file. In the
case of a TxDb object, this is obtained by \code{GenomicFeatures::exonBy}.
\item \code{intronic_parts}: a GRangesList column containing the intronic regions
for each gene. This is simply defined as the parallel set difference betweeen
the gene ranges and the exonic ranges.
\item \code{n_olaps}: the integer count of the number times a gene overlaps
any other gene in the annotation.
}
}
\examples{
gtf <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf", package = "airway")
collect_parts(gtf)

}
