% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_rules}
\alias{filter_rules}
\title{Recommended filters for analysis}
\usage{
filter_rules(.data, ...)
}
\arguments{
\item{.data}{A GRanges object obtained from `prepare_annotation()`}

\item{...}{A set of logical expressions to filter `.data` if 
no expressions are provided the default rules are applied.}
}
\description{
Recommended filters for analysis
}
\details{
This runs the default set of filters before adding coverage
information, which are:

* Protein coding genes only
* Does not overlap any other genes
* Greater than 1 exonic feature
}
