% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_parts}
\alias{join_parts}
\title{Expand exonic/intronic parts and intersect with ranges}
\usage{
join_parts(x, parts)
}
\arguments{
\item{x}{a GRanges object}

\item{parts}{an annotation GRanges from \code{collect_parts()}}
}
\value{
a GRanges object
}
\description{
Expand exonic/intronic parts and intersect with ranges
}
\details{
This is a wrapper to \code{unnest_parts()} and
\code{plyranges::join_overlap_intersect()}. It restricts the GRanges object \code{x}
to the ranges that intersect the \code{exonic_parts} and \code{intronic_parts} columns
from \code{parts}, the result is a GRanges object with additional columns:
\itemize{
\item \code{gene_id}, the gene_id from \code{features} corresponding to an exon/intron.
\item \code{feature_type}, whether the range corresponds to an exon or intro feature.
\item \code{feature_rank}, the rank of the exon/intron feature within a gene.
\item \code{feature_length}, the width of the exon/intron
\item \code{feature_strand}, the strand of the exon/intron
}
}
