% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{nest_by_overlaps}
\alias{nest_by_overlaps}
\title{Nest by Overlapping Ranges}
\usage{
nest_by_overlaps(ranges, features, .key, .index)
}
\arguments{
\item{ranges}{a GRanges object}

\item{features}{a GRanges object from \code{collect_parts()}}

\item{.key}{one or more columns to group ranges}

\item{.index}{one or more columns to group overlaps by}
}
\value{
a GRanges object
}
\description{
Nest by Overlapping Ranges
}
\details{
This function restricts the coverage GRanges, \code{cvg} to the ranges
that intersect the \code{exonic_parts} and \code{intronic_parts} ranges, the result
is a GRanges object with additional columns:
\itemize{
\item \code{gene_id}, the gene_id from \code{features} corresponding to an exon/intron.
\item \code{feature_type}, whether the range corresponds to an exon or intro feature.
\item \code{feature_rank}, the rank of the exon/intron feature within a gene.
\item \code{feature_length}, the width of the exon/intron
\item \code{feature_strand}, the strand of the exon/intron
}
}
