% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_annotation}
\alias{prepare_annotation}
\title{Prepare GTF/GFF file for coverage analysis by
splitting into intron/exon features}
\usage{
prepare_annotation(.data)
}
\arguments{
\item{.data}{a GRanges object obtained from `rtracklayer::import()` or
`plyranges::read_gff()`.}
}
\description{
Prepare GTF/GFF file for coverage analysis by
splitting into intron/exon features
}
