% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangenostics.R
\docType{methods}
\name{rangle}
\alias{rangle}
\alias{rangle,GenomicRanges-method}
\alias{rangle,GroupedGenomicRanges-method}
\title{Compute diagnostics over Ranges}
\usage{
rangle(x, .var, .index, .funs)

\S4method{rangle}{GenomicRanges}(x, .var, .index, .funs)

\S4method{rangle}{GroupedGenomicRanges}(x, .var, .index, .funs)
}
\arguments{
\item{x}{a Ranges object}

\item{.var}{A metadata column in \code{x} to summarise over either a bare variable name
or a character vector of length 1.}

\item{.index}{Bare variable names that index each Range}

\item{.funs}{A named-list of functions that compute diagnostics.}
}
\description{
Compute diagnostics over Ranges
}
\examples{
gr <- GRanges(
seqnames = factor(sample(lvls, 1000, replace = TRUE), levels = lvls),
ranges = IRanges(start = rpois(1000, 10000), width = rpois(1000, 100)),
grp = sample(letters[1:4], 1000, replace = TRUE),
score = rnbinom(1000, size = 5, mu = 25)
)

rangle(gr, score, seqnames, list(mean = mean))


}
