% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangle.R
\docType{methods}
\name{rangle}
\alias{rangle}
\alias{rangle,GenomicRanges-method}
\alias{rangle,GroupedGenomicRanges-method}
\title{Rangle new features out of .ranges}
\usage{
rangle(x, .var, .index, .funs)

\S4method{rangle}{GenomicRanges}(x, .var, .index, .funs)

\S4method{rangle}{GroupedGenomicRanges}(x, .var, .index, .funs)
}
\arguments{
\item{x}{a Ranges object}

\item{.var}{A metadata column in \code{x} to summarise over either a bare variable name
or a character vector of length 1.}

\item{.index}{A list of columns generated with \code{rng_vars()}
that index the ranges in \code{x}.}

\item{.funs}{A named-list of functions that compute \code{rangenostics()}}
}
\description{
Rangle new features out of .ranges
}
\examples{
suppressPackageStartupMessages(library("GenomicRanges"))
lvls <- paste0("chr", 1:23)
N <- 1e5
gr <- GRanges(
seqnames = factor(sample(lvls, N, replace = TRUE), levels = lvls),
ranges = IRanges(start = rpois(N, 10000), width = rpois(N, 100)),
grp = sample(letters[1:4], N, replace = TRUE),
score = rnbinom(N, size = 5, mu = 25)
)

rangle(gr, score, rng_vars(seqnames), list(mean = mean))


}
