% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangenostics.R
\name{lumpy}
\alias{lumpy}
\alias{bumpy}
\alias{sum_above}
\alias{flat_spots}
\alias{count_grams}
\alias{max_mean_shift}
\alias{max_var_shift}
\alias{gammy}
\title{Rangewise diagnostics}
\usage{
lumpy(x, .size)

bumpy(x, .size)

sum_above(x, .threshold)

flat_spots(x, .threshold = 0, .tol = .Machine$double.eps^0.5)

count_grams(x, .size)

max_mean_shift(x, .size, .step)

max_var_shift(x, .size, .step)

gammy(x)
}
\arguments{
\item{x}{a numeric vector or Rle object}

\item{.size}{the integer size of the window}
}
\description{
Rangewise diagnostics
}
\details{
These are inspired by the concept of cognostics by John Tukey, and related work in the
area by Lee Wilkinson (graph theoretic scagnostics) and in time series analysis by Rob Hyndman.

The
}
