% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{spread_coverage}
\alias{spread_coverage}
\title{Spread intron/exon coverage over a gene}
\usage{
spread_coverage(x, ..., threshold = 0L)
}
\arguments{
\item{x}{a GRanges object from `merge_coverage()\merge_design()`}

\item{...}{columns to partition x}

\item{threshold}{Filter bases covered to above some integer threshold (default = 0L).}
}
\description{
Spread intron/exon coverage over a gene
}
\details{
This function summarises the *total* coverage within an exon/intron
for each gene, it then widens the result so there are seperate columns
for each exon/intron variable.
}
