% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{squish_coverage}
\alias{squish_coverage}
\title{Squish coverage over intronic/exonic parts}
\usage{
squish_coverage(x, ..., threshold = 0L)
}
\arguments{
\item{x}{a GRanges object from \code{merge_coverage()} and or \code{merge_design()}}

\item{...}{columns to partition x}

\item{threshold}{Filter bases covered to above some integer threshold (default = 0L).}
}
\value{
a GRanges object, where each gene is split into exonic
and intronic parts, and the coverage is summarised as follows over each
exon / intron:
\itemize{
\item \code{gene_id} the gene_id as from \code{prepare_annotation()}
\item \code{feature_type} whether a feature is an intron or exon
\item \code{feature_length} the total length of the intronic/exonic part
\item \code{total_coverage}
\item \code{average_coverage}
\item \code{n_bases_above_threshold}
\item \code{prop_bases_covered_at_threshold}
}
}
\description{
Squish coverage over intronic/exonic parts
}
\details{
Here we use \code{plyranges::reduce_ranges()} to squish
the coverage GRanges to have coverage over each exon/intron within
a gene. If \code{...} are supplied additional grouping is added by metadata
columns in x, using \code{plyranges::group_by()}.
}
