% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-coverage.R
\name{view_coverage}
\alias{view_coverage}
\title{View coverage as a stacked area chart}
\usage{
view_coverage(cvg, target, hline = NULL, facets = ggplot2::vars())
}
\arguments{
\item{cvg}{a GRanges object obtained from \code{merge_coverage()}}

\item{target}{a GRanges object from \code{prepare_annotation()} containing a gene_id column}

\item{hline}{an optional data.frame giving marker points}

\item{facets}{a set of variables quoted by \code{ggplot2::vars()} that is passed
down to \code{facet_wrap()}}
}
\description{
View coverage as a stacked area chart
}
