% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-coverage.R
\name{view_coverage}
\alias{view_coverage}
\title{View coverage as an area chart oriented from 5' to 3'}
\usage{
view_coverage(data, score, colour = NULL, facets = rng_vars())
}
\arguments{
\item{data}{a GRanges object}

\item{score}{an expression or column in data that represents the coverage score}

\item{colour}{a column in the data to highlight ranges by}

\item{facets}{a set of metadata or range variables in \code{data} quoted by \code{rng_vars()} that is passed
down to \code{facet_wrap()}}
}
\description{
View coverage as an area chart oriented from 5' to 3'
}
