% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{tile_rle}
\alias{tile_rle}
\alias{stretch_rle}
\alias{roll_rle}
\title{Map a function over an Rle in windows}
\usage{
tile_rle(.x, .size = 1L, .fun, ...)

stretch_rle(.x, .size = 1L, .step = 1L, .fun, ...)

roll_rle(.x, .size = 1L, .step = 1L, .fun, ...)
}
\arguments{
\item{.x}{An atomic vector or Rle object.}

\item{.size}{The (integer) size of the window.}

\item{.fun}{A function}

\item{...}{Additional arguments passed on to the mapped function}

\item{.step}{The (integer) amount to shift the start of the window by.}
}
\description{
Map a function over an Rle in windows
}
\details{
The map functions apply a function over non-overlapping windows \verb{[tile_rle()]},
overlapping windows \verb{[roll_rle()]}, and windows with a fixed start but
increasing width \code{\link[=stretch_rle]{stretch_rle()}}
}
\examples{
x <- S4Vectors::Rle(1:10, 1:10)
tile_rle(x, .size = 2, mean)
roll_rle(x, .size = 2, .step = 1, mean)
stretch_rle(x, .size = 1, .step = 2, mean)


}
